chore: merged getQuote and getRoute scripts into `yarn quote`
