+++ .vscode/settings.json
@@ -1,3 +1,12 @@
 {
-  "cSpell.words": ["ceur", "ckes", "creal", "cusd", "EUROC", "PUSO"]
+  "cSpell.words": [
+    "ceur",
+    "cfgs",
+    "ckes",
+    "creal",
+    "cusd",
+    "EUROC",
+    "Numberish",
+    "PUSO"
+  ]
 }
