--- scripts/getRoute.ts
@@ -1,7 +1,0 @@
-#!/usr/bin/env ts-node
-
-/**
- * This is a wrapper script that uses the modular implementation.
- * For better maintainability, the implementation has been moved to the poolConfig folder.
- */
-import './getRoute/index'
